Add a CI build in Github workflow, transition to vcpkg (#5)
Implements an actual CI build in a Github workflow.   Though it started out innocently enough, dependency resolution quickly got out of hand (*cough* Pangolin).  This led to a wholesale conversion to `vcpkg` for standalone, non-ROS builds, both in and out of CI.

README updated to match.

+++ src/Settings.cc
@@ -117,7 +117,7 @@ void Settings::setMonoCamera(CameraType type, const std::vector<float>& k,
     //       vOverlapping;
     // }
   } else {
-    spdlog::error("Error: {} not known", type);
+    spdlog::error("Error: {} not known", static_cast<int>(type));
     exit(-1);
   }
 }
